refactor: 전역 함수 및 전역 변수를 지역 함수, 변수로 변경 makeRandomNumberOfCustomers()와 numberOfCustomers를 Bank 구조체 안으로 넣어줌
